added stuff

+++ app/src/main/java/com/linute/linute/MainContent/Settings/ChangeProfileImageFragment.java
@@ -3,23 +3,30 @@ package com.linute.linute.MainContent.Settings;
 import android.Manifest;
 import android.animation.Animator;
 import android.animation.AnimatorListenerAdapter;
+import android.annotation.TargetApi;
+import android.app.Activity;
 import android.app.AlertDialog;
+import android.app.Fragment;
+import android.content.Context;
 import android.content.DialogInterface;
 import android.content.Intent;
 import android.content.SharedPreferences;
 import android.content.pm.PackageManager;
 import android.graphics.Bitmap;
 import android.net.Uri;
+import android.os.Build;
 import android.os.Environment;
 import android.provider.MediaStore;
+import android.support.annotation.Nullable;
 import android.support.v4.app.ActivityCompat;
 import android.support.v4.content.ContextCompat;
-import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 import android.util.Log;
+import android.view.LayoutInflater;
 import android.view.View;
-import android.widget.Button;
+import android.view.ViewGroup;
 import android.widget.ProgressBar;
+import android.widget.TextView;
 import android.widget.Toast;
 
 import com.bumptech.glide.Glide;
@@ -27,10 +34,10 @@ import com.bumptech.glide.load.engine.DiskCacheStrategy;
 import com.bumptech.glide.signature.StringSignature;
 import com.linute.linute.API.LSDKUser;
 import com.linute.linute.R;
-import com.linute.linute.SquareCamera.SquareImageView;
 import com.linute.linute.UtilsAndHelpers.LinuteConstants;
 import com.linute.linute.UtilsAndHelpers.LinuteUser;
 import com.linute.linute.UtilsAndHelpers.Utils;
+import com.mikhaellopez.circularimageview.CircularImageView;
 import com.soundcloud.android.crop.Crop;
 import com.squareup.okhttp.Callback;
 import com.squareup.okhttp.Request;
@@ -49,19 +56,17 @@ import java.util.List;
 import java.util.Map;
 import java.util.Random;
 
-public class ChangeProfileImageActivity extends AppCompatActivity {
+public class ChangeProfileImageFragment extends Fragment {
 
     public static final String TAG = "ChangeProfileImage";
 
-    private boolean mHasSavedImage = false; //if image was successfully saved to DB, update profile frag
-
     private boolean mHasChangedImage = false; //won't allow send unless user actually makes a change
 
-    private Button mSaveButton;
-    private Button mCancelButton;
+    private TextView mEditButton;
+    private TextView mSaveButton;
     private View mButtonLayer;
 
-    private SquareImageView mImageView;
+    private CircularImageView mImageView;
 
     private SharedPreferences mSharedPreferences;
 
@@ -69,63 +74,50 @@ public class ChangeProfileImageActivity extends AppCompatActivity {
 
     private Bitmap mProfilePictureBitmap;
 
-
+    @Nullable
     @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_change_profile_image);
-        mSharedPreferences = getSharedPreferences(LinuteConstants.SHARED_PREF_NAME, MODE_PRIVATE);
+    public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
+        View rootView = inflater.inflate(R.layout.fragment_change_profile_image, container, false);
+        mSharedPreferences = getActivity().getSharedPreferences(LinuteConstants.SHARED_PREF_NAME, Context.MODE_PRIVATE);
 
-        bindViews();
-
-        if (!Utils.isNetworkAvailable(this)) Utils.showBadConnectionToast(this);
+        bindViews(rootView);
         setDefaultValues();
-
         setUpOnClickListeners();
+
+        ((EditProfileInfoActivity)getActivity()).setTitle("Photo");
+
+        return rootView;
     }
 
+    private void bindViews(View rootView) {
+        mSaveButton = (TextView) rootView.findViewById(R.id.changeprofileimage_save_button);
+        mEditButton = (TextView) rootView.findViewById(R.id.changeprofileimage_change_button);
+        mButtonLayer = rootView.findViewById(R.id.changeprofileimage_buttons);
 
-    private void bindViews() {
-        mSaveButton = (Button) findViewById(R.id.changeprofileimage_save_button);
-        mCancelButton = (Button) findViewById(R.id.changeprofileimage_cancel_button);
-        mButtonLayer = findViewById(R.id.changeprofileimage_buttons);
+        mImageView = (CircularImageView) rootView.findViewById(R.id.changeprofileimage_image);
 
-        mImageView = (SquareImageView) findViewById(R.id.changeprofileimage_image);
-
-        mProgressBar = (ProgressBar) findViewById(R.id.changeprofileimage_progressbar);
+        mProgressBar = (ProgressBar) rootView.findViewById(R.id.changeprofileimage_progressbar);
     }
 
     private void setUpOnClickListeners() {
-        mSaveButton.setOnClickListener(new View.OnClickListener() {
+        mEditButton.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                saveImage();
-            }
-        });
-        mCancelButton.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                onBackPressed();
-                overridePendingTransition(0, 0);
-            }
-        });
-
-        mImageView.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                AlertDialog.Builder builder = new AlertDialog.Builder(ChangeProfileImageActivity.this);
+                AlertDialog.Builder builder = new AlertDialog.Builder(getActivity());
                 builder.setTitle("Image Source");
                 String[] options = {"Camera", "Photo Gallery", "Cancel"};
                 builder.setItems(options, actionListener);
                 builder.create().show();
             }
         });
+        mSaveButton.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                saveImage();
+            }
+        });
     }
 
-    @Override
-    protected void onResume() {
-        super.onResume();
-    }
 
     //select between camera and photogallery
     DialogInterface.OnClickListener actionListener = new DialogInterface.OnClickListener() {
@@ -138,7 +130,7 @@ public class ChangeProfileImageActivity extends AppCompatActivity {
                     break;
                 case 1:
                     //go to gallery
-                    Crop.pickImage(ChangeProfileImageActivity.this);
+                    Crop.pickImage(getActivity(), ChangeProfileImageFragment.this);
                     break;
                 case 2:
                     break;
@@ -148,19 +140,8 @@ public class ChangeProfileImageActivity extends AppCompatActivity {
         }
     };
 
-    @Override
-    public void onBackPressed() {
-        if (mProgressBar.getVisibility() == View.GONE) {
-            if (mHasSavedImage) setResult(RESULT_OK); //tell parent to update
-            else setResult(RESULT_CANCELED); //tell parent not to update
-            super.onBackPressed();
-        }
-    }
-
-
 
     private void setDefaultValues() {
-
         Glide.with(this)
                 .load(Utils.getImageUrlOfUser(mSharedPreferences.getString("profileImage", "")))
                 .asBitmap()
@@ -173,7 +154,7 @@ public class ChangeProfileImageActivity extends AppCompatActivity {
 
     private void saveImage() {
         if (!mHasChangedImage) return; //no edits to image
-        LSDKUser user = new LSDKUser(this);
+        LSDKUser user = new LSDKUser(getActivity());
         showProgress(true);
 
         Map<String, Object> userInfo = new HashMap<>();
@@ -182,10 +163,11 @@ public class ChangeProfileImageActivity extends AppCompatActivity {
         user.updateUserInfo(userInfo, null, new Callback() {
             @Override
             public void onFailure(Request request, IOException e) {
-                runOnUiThread(new Runnable() {
+                if (getActivity() == null) return;
+                getActivity().runOnUiThread(new Runnable() {
                     @Override
                     public void run() {
-                        Utils.showBadConnectionToast(ChangeProfileImageActivity.this);
+                        Utils.showBadConnectionToast(getActivity());
                         showProgress(false);
                     }
                 });
@@ -197,39 +179,43 @@ public class ChangeProfileImageActivity extends AppCompatActivity {
                     try {
                         LinuteUser user = new LinuteUser(new JSONObject(response.body().string()));
                         persistData(user);
-                        mHasSavedImage = true;
+
                         mHasChangedImage = false;
 
-                        runOnUiThread(new Runnable() {
+                        final EditProfileInfoActivity activity = (EditProfileInfoActivity) getActivity();
+                        if (activity == null) return;
+                        activity.runOnUiThread(new Runnable() {
                             @Override
                             public void run() {
-                                Utils.showSavedToast(ChangeProfileImageActivity.this);
+                                Utils.showSavedToast(activity);
+                                showProgress(false);
+
+                                activity.setMainActivityNeedsToUpdate(true);
+                                getFragmentManager().popBackStack();
                             }
                         });
                     } catch (JSONException e) {
                         e.printStackTrace();
-                        runOnUiThread(new Runnable() {
+                        if (getActivity() == null) return;
+                        getActivity().runOnUiThread(new Runnable() {
                             @Override
                             public void run() {
-                                Utils.showServerErrorToast(ChangeProfileImageActivity.this);
+                                Utils.showServerErrorToast(getActivity());
+                                showProgress(false);
                             }
                         });
                     }
                 } else {
                     Log.v(TAG, response.body().string());
-                    runOnUiThread(new Runnable() {
+                    if (getActivity() == null) return;
+                    getActivity().runOnUiThread(new Runnable() {
                         @Override
                         public void run() {
-                            Utils.showServerErrorToast(ChangeProfileImageActivity.this);
+                            Utils.showServerErrorToast(getActivity());
+                            showProgress(false);
                         }
                     });
                 }
-                runOnUiThread(new Runnable() {
-                    @Override
-                    public void run() {
-                        showProgress(false);
-                    }
-                });
             }
         });
     }
@@ -238,36 +224,46 @@ public class ChangeProfileImageActivity extends AppCompatActivity {
     static final int REQUEST_TAKE_PHOTO = 1;
     private String mCurrentPhotoPath;
 
+
+
     @Override
-    protected void onActivityResult(int requestCode, int resultCode, Intent data) {
-        super.onActivityResult(requestCode, resultCode, data);
+    public void onActivityResult(int requestCode, int resultCode, Intent data) {
+
+        if (mCurrentPhotoPath == null) return;
+
+        if (!hasWritePermission() && !hasCameraPermissions()){
+            if (mCurrentPhotoPath != null){
+                new File(mCurrentPhotoPath).delete();
+                mCurrentPhotoPath = null;
+            }
+            return;
+        }
 
         if (requestCode == REQUEST_TAKE_PHOTO) { //got response from camera
-            if (resultCode == RESULT_OK) {  //was able to get picture
-                if (hasWritePermission()) {
-                    File f = new File(mCurrentPhotoPath);
-                    Uri contentUri = Uri.fromFile(f);
-                    galleryAddPic(contentUri); // add to gallery
-                    beginCrop(contentUri); //crop image
-                }else
-                    showRationalizationDialog();
-            } else { //no picture captured. delete the temp file created to hold image
+            if (resultCode == Activity.RESULT_OK) {  //was able to get picture
+                File f = new File(mCurrentPhotoPath);
+                Uri contentUri = Uri.fromFile(f);
+                galleryAddPic(contentUri); // add to gallery
+                beginCrop(contentUri); //crop image
+            }
+             else { //no picture captured. delete the temp file created to hold image
                 if (!new File(mCurrentPhotoPath).delete())
                     Log.v(TAG, "could not delete temp file");
                 mCurrentPhotoPath = null;
             }
-        } else if (requestCode == Crop.REQUEST_PICK && resultCode == RESULT_OK) { //got image from gallery
+        } else if (requestCode == Crop.REQUEST_PICK && resultCode == Activity.RESULT_OK) { //got image from gallery
             beginCrop(data.getData()); //crop image
         } else if (requestCode == Crop.REQUEST_CROP) { //photo came back from crop
-            if (resultCode == RESULT_OK) {
+            if (resultCode == Activity.RESULT_OK) {
                 Uri imageUri = Crop.getOutput(data);
                 try {
                     //release old pictures resources
                     if (mProfilePictureBitmap != null) mProfilePictureBitmap.recycle();
 
+
                     //scale cropped image to 1080 x 1080 (will be sent to database
                     mProfilePictureBitmap = Bitmap.createScaledBitmap(
-                            MediaStore.Images.Media.getBitmap(this.getContentResolver(), imageUri),
+                            MediaStore.Images.Media.getBitmap(getActivity().getContentResolver(), imageUri),
                             1080, 1080, false);
 
                     mImageView.setImageBitmap(mProfilePictureBitmap);
@@ -277,15 +273,15 @@ public class ChangeProfileImageActivity extends AppCompatActivity {
                     e.printStackTrace();
                 }
             } else if (resultCode == Crop.RESULT_ERROR) { //error cropping, show error
-                Toast.makeText(this, Crop.getError(data).getMessage(), Toast.LENGTH_SHORT).show();
+                Toast.makeText(getActivity(), Crop.getError(data).getMessage(), Toast.LENGTH_SHORT).show();
             }
         }
     }
 
 
     private void beginCrop(Uri source) { //begin crop activity
-        Uri destination = Uri.fromFile(new File(getCacheDir(), "cropped"));
-        Crop.of(source, destination).asSquare().start(this);
+        Uri destination = Uri.fromFile(new File(getActivity().getCacheDir(), "cropped"));
+        Crop.of(source, destination).asSquare().start(getActivity(), ChangeProfileImageFragment.this);
     }
 
 
@@ -326,6 +322,7 @@ public class ChangeProfileImageActivity extends AppCompatActivity {
     //Request Permissiosns
     private static final int REQUEST_PERMISSIONS = 10;
 
+    @TargetApi(Build.VERSION_CODES.M)
     public void requestPermissions(){
         List<String> permissions = new ArrayList<>();
         //check for camera
@@ -338,9 +335,8 @@ public class ChangeProfileImageActivity extends AppCompatActivity {
         }
         //we need permissions
         if (!permissions.isEmpty()){
-            ActivityCompat.requestPermissions(this,
-                    permissions.toArray(new String[permissions.size()]),
-                    REQUEST_PERMISSIONS);
+           requestPermissions(permissions.toArray(new String[permissions.size()]),
+                   REQUEST_PERMISSIONS);
         }else {
             //we have permissions : show camera
             dispatchTakePictureIntent();
@@ -349,6 +345,7 @@ public class ChangeProfileImageActivity extends AppCompatActivity {
 
     @Override
     public void onRequestPermissionsResult(int requestCode, String[] permissions, int[] grantResults) {
+        Log.i(TAG, "onRequestPermissionsResult: ");
         switch (requestCode) {
             case REQUEST_PERMISSIONS:
                 for (int result : grantResults) // if we didn't get approved for a permission, show permission needed frag
@@ -364,9 +361,9 @@ public class ChangeProfileImageActivity extends AppCompatActivity {
     }
 
     private void showRationalizationDialog() {
-        new AlertDialog.Builder(this)
-                .setTitle("Allow Woohoo to Use your phone's storage?")
-                .setMessage("Woohoo needs access to your phone's camera and storage to take and save images.")
+        new AlertDialog.Builder(getActivity())
+                .setTitle("Allow Tapt to Use your phone's storage?")
+                .setMessage("Tapt needs access to your phone's camera and storage to take and save images.")
                 .setPositiveButton("Allow", new DialogInterface.OnClickListener() {
                     @Override
                     public void onClick(DialogInterface dialog, int which) {
@@ -386,7 +383,7 @@ public class ChangeProfileImageActivity extends AppCompatActivity {
     private void dispatchTakePictureIntent() {
         Intent takePictureIntent = new Intent(MediaStore.ACTION_IMAGE_CAPTURE);
         // Ensure that there's a camera activity to handle the intent
-        if (takePictureIntent.resolveActivity(getPackageManager()) != null) {
+        if (takePictureIntent.resolveActivity(getActivity().getPackageManager()) != null) {
             // Create the File where the photo should go
             File photoFile = null;
             try {
@@ -410,7 +407,7 @@ public class ChangeProfileImageActivity extends AppCompatActivity {
         String imageFileName = "JPEG_" + timeStamp + "_";
         //create folder for our pictures
         File storageDir = new File(Environment.getExternalStoragePublicDirectory(
-                Environment.DIRECTORY_PICTURES), "Linute");
+                Environment.DIRECTORY_PICTURES), "Tapt");
 
         if (!storageDir.exists()) storageDir.mkdir();
 
@@ -429,16 +426,16 @@ public class ChangeProfileImageActivity extends AppCompatActivity {
     private void galleryAddPic(Uri contentUri) {
         Intent mediaScanIntent = new Intent(Intent.ACTION_MEDIA_SCANNER_SCAN_FILE);
         mediaScanIntent.setData(contentUri);
-        this.sendBroadcast(mediaScanIntent);
+        getActivity().sendBroadcast(mediaScanIntent);
     }
 
 
     private boolean hasWritePermission(){
-        return ContextCompat.checkSelfPermission(this, Manifest.permission.WRITE_EXTERNAL_STORAGE) == PackageManager.PERMISSION_GRANTED;
+        return ContextCompat.checkSelfPermission(getActivity(), Manifest.permission.WRITE_EXTERNAL_STORAGE) == PackageManager.PERMISSION_GRANTED;
     }
 
     private boolean hasCameraPermissions(){
-        return ContextCompat.checkSelfPermission(this, Manifest.permission.CAMERA) == PackageManager.PERMISSION_GRANTED;
+        return ContextCompat.checkSelfPermission(getActivity(), Manifest.permission.CAMERA) == PackageManager.PERMISSION_GRANTED;
 
     }
 }
